Move tests to seperate package and update GitHub Actions accordingly
