+++ .github/workflows/e2e-tests.yaml
@@ -42,8 +42,11 @@ jobs:
           FAUCET_PUBLIC_KEY: 782127638ab9cc8fa8598ff0120464ceef6f367ddcd616b47afffbdb175dcc5e
           FAUCET_PRIVATE_KEY: 93fd78bd1e79a3b593d4732b76050b418aeefa5e4e6ea80f351b4c3a5b0dd19f
 
+      - name: Start Tools dApp
+        run: pnpm --filter @kadena/tools start
+
       - name: Run Playwright tests
-        run: pnpm run test:e2e --filter @kadena/tools
+        run: pnpm run test:e2e --project=tools-app
         env:
           CI: true
