+++ README.md
@@ -26,18 +26,19 @@ Overview of the main packages maintained in this repository:
 
 | Package                              | Release Notes        |
 | :----------------------------------- | :------------------- |
-| [@kadena-dev/eslint-config][5]       | [![version][7]][6]   |
-| [@kadena-dev/eslint-plugin][8]       | [![version][10]][9]  |
-| [@kadena-dev/heft-rig][11]           | [![version][13]][12] |
-| [@kadena/chainweb-node-client][14]   | [![version][16]][15] |
-| [@kadena/chainweb-stream-client][17] | [![version][19]][18] |
-| [@kadena/chainwebjs][20]             | [![version][22]][21] |
-| [@kadena/client][23]                 | [![version][25]][24] |
-| [@kadena/cryptography-utils][26]     | [![version][28]][27] |
-| [@kadena/pactjs][29]                 | [![version][31]][30] |
-| [@kadena/pactjs-cli][32]             | [![version][34]][33] |
-| [@kadena/pactjs-generator][35]       | [![version][37]][36] |
-| [@kadena/types][38]                  | [![version][40]][39] |
+| [@kadena-dev/e2e-tests][5]           | [![version][7]][6]   |
+| [@kadena-dev/eslint-config][8]       | [![version][10]][9]  |
+| [@kadena-dev/eslint-plugin][11]      | [![version][13]][12] |
+| [@kadena-dev/heft-rig][14]           | [![version][16]][15] |
+| [@kadena/chainweb-node-client][17]   | [![version][19]][18] |
+| [@kadena/chainweb-stream-client][20] | [![version][22]][21] |
+| [@kadena/chainwebjs][23]             | [![version][25]][24] |
+| [@kadena/client][26]                 | [![version][28]][27] |
+| [@kadena/cryptography-utils][29]     | [![version][31]][30] |
+| [@kadena/pactjs][32]                 | [![version][34]][33] |
+| [@kadena/pactjs-cli][35]             | [![version][37]][36] |
+| [@kadena/pactjs-generator][38]       | [![version][40]][39] |
+| [@kadena/types][41]                  | [![version][43]][42] |
 
 <!--packageTable end -->
 
@@ -45,59 +46,63 @@ Overview of the main packages maintained in this repository:
 
 Special thanks to the wonderful people who have contributed to this project:
 
-[![Contributors][42]][41]
+[![Contributors][45]][44]
 
 [1]: https://docs.kadena.io
 [2]: https://discord.io/kadena
 [3]: https://stackoverflow.com/questions/tagged/kadena
 [4]: ./CONTRIBUTING.md
 [5]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/tools/eslint-config
-[6]: packages/tools/eslint-config/CHANGELOG.md
-[7]: https://img.shields.io/npm/v/@kadena-dev/eslint-config.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/tools/e2e-tests
+[6]: packages/tools/e2e-tests/CHANGELOG.md
+[7]: https://img.shields.io/npm/v/@kadena-dev/e2e-tests.svg
 [8]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/tools/eslint-plugin
-[9]: packages/tools/eslint-plugin/CHANGELOG.md
-[10]: https://img.shields.io/npm/v/@kadena-dev/eslint-plugin.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/tools/eslint-config
+[9]: packages/tools/eslint-config/CHANGELOG.md
+[10]: https://img.shields.io/npm/v/@kadena-dev/eslint-config.svg
 [11]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/tools/heft-rig
-[12]: packages/tools/heft-rig/CHANGELOG.md
-[13]: https://img.shields.io/npm/v/@kadena-dev/heft-rig.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/tools/eslint-plugin
+[12]: packages/tools/eslint-plugin/CHANGELOG.md
+[13]: https://img.shields.io/npm/v/@kadena-dev/eslint-plugin.svg
 [14]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/chainweb-node-client
-[15]: packages/libs/chainweb-node-client/CHANGELOG.md
-[16]: https://img.shields.io/npm/v/@kadena/chainweb-node-client.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/tools/heft-rig
+[15]: packages/tools/heft-rig/CHANGELOG.md
+[16]: https://img.shields.io/npm/v/@kadena-dev/heft-rig.svg
 [17]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/chainweb-stream-client
-[18]: packages/libs/chainweb-stream-client/CHANGELOG.md
-[19]: https://img.shields.io/npm/v/@kadena/chainweb-stream-client.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/chainweb-node-client
+[18]: packages/libs/chainweb-node-client/CHANGELOG.md
+[19]: https://img.shields.io/npm/v/@kadena/chainweb-node-client.svg
 [20]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/chainwebjs
-[21]: packages/libs/chainwebjs/CHANGELOG.md
-[22]: https://img.shields.io/npm/v/@kadena/chainwebjs.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/chainweb-stream-client
+[21]: packages/libs/chainweb-stream-client/CHANGELOG.md
+[22]: https://img.shields.io/npm/v/@kadena/chainweb-stream-client.svg
 [23]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/client
-[24]: packages/libs/client/CHANGELOG.md
-[25]: https://img.shields.io/npm/v/@kadena/client.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/chainwebjs
+[24]: packages/libs/chainwebjs/CHANGELOG.md
+[25]: https://img.shields.io/npm/v/@kadena/chainwebjs.svg
 [26]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/cryptography-utils
-[27]: packages/libs/cryptography-utils/CHANGELOG.md
-[28]: https://img.shields.io/npm/v/@kadena/cryptography-utils.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/client
+[27]: packages/libs/client/CHANGELOG.md
+[28]: https://img.shields.io/npm/v/@kadena/client.svg
 [29]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/pactjs
-[30]: packages/libs/pactjs/CHANGELOG.md
-[31]: https://img.shields.io/npm/v/@kadena/pactjs.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/cryptography-utils
+[30]: packages/libs/cryptography-utils/CHANGELOG.md
+[31]: https://img.shields.io/npm/v/@kadena/cryptography-utils.svg
 [32]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/tools/pactjs-cli
-[33]: packages/tools/pactjs-cli/CHANGELOG.md
-[34]: https://img.shields.io/npm/v/@kadena/pactjs-cli.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/pactjs
+[33]: packages/libs/pactjs/CHANGELOG.md
+[34]: https://img.shields.io/npm/v/@kadena/pactjs.svg
 [35]:
-  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/pactjs-generator
-[36]: packages/libs/pactjs-generator/CHANGELOG.md
-[37]: https://img.shields.io/npm/v/@kadena/pactjs-generator.svg
+  https://github.com/kadena-community/kadena.js/tree/main/packages/tools/pactjs-cli
+[36]: packages/tools/pactjs-cli/CHANGELOG.md
+[37]: https://img.shields.io/npm/v/@kadena/pactjs-cli.svg
 [38]:
+  https://github.com/kadena-community/kadena.js/tree/main/packages/libs/pactjs-generator
+[39]: packages/libs/pactjs-generator/CHANGELOG.md
+[40]: https://img.shields.io/npm/v/@kadena/pactjs-generator.svg
+[41]:
   https://github.com/kadena-community/kadena.js/tree/main/packages/libs/types
-[39]: packages/libs/types/CHANGELOG.md
-[40]: https://img.shields.io/npm/v/@kadena/types.svg
-[41]: https://github.com/kadena-community/kadena.js/graphs/contributors
-[42]: https://contrib.rocks/image?repo=kadena-community/kadena.js
+[42]: packages/libs/types/CHANGELOG.md
+[43]: https://img.shields.io/npm/v/@kadena/types.svg
+[44]: https://github.com/kadena-community/kadena.js/graphs/contributors
+[45]: https://contrib.rocks/image?repo=kadena-community/kadena.js
