--- packages/apps/tools/e2e-tests/fixtures/mocks/send.ts
@@ -1,3 +1,0 @@
-export const send = {
-  requestKeys: ['rNk4ePNVlJGwjCaVrjKFJbh3A0agpcZq6gYLfivOHYg'],
-};
