--- packages/apps/tools/e2e-tests/tscconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "./node_modules/@kadena-dev/heft-rig/tsconfig-playwright.json"
-}
