--- packages/tools/e2e-components/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@kadena/e2e-components",
-  "version": "1.0.0",
-  "description": "",
-  "keywords": [],
-  "license": "ISC",
-  "author": "",
-  "scripts": {
-  },
-  "devDependencies": {
-    "@playwright/test": "^1.38.1"
-  }
-}
