+++ packages/tools/e2e-tests/README.MD
@@ -1,0 +1,15 @@
+<!-- genericHeader start -->
+
+# @kadena/client
+
+Core library for building Pact expressions to send to the blockchain in js.
+Makes use of .kadena/pactjs-generated
+
+<picture>
+  <source srcset="https://raw.githubusercontent.com/kadena-community/kadena.js/main/common/images/Kadena.JS_logo-white.png" media="(prefers-color-scheme: dark)"/>
+  <img src="https://raw.githubusercontent.com/kadena-community/kadena.js/main/common/images/Kadena.JS_logo-black.png" width="200" alt="kadena.js logo" />
+</picture>
+
+<!-- genericHeader end -->
+
+TODO
