+++ packages/tools/e2e-tests/fixtures/data/accounts.ts
@@ -1,0 +1,8 @@
+interface IAccounts {
+  publicKey: string;
+  secretkey: string;
+}
+export const accountsData: IAccounts = {
+  publicKey: '938c8ec66283493844c205481f220cfbd7024bda5859c7897141f6a5c6fc52c1',
+  secretkey: 'be6442bd49bb7c1c48413596da6e879f85c047bbe8236abe9f9afb63eb32705d',
+};
