+++ packages/tools/e2e-tests/fixtures/mocks/poll.ts
@@ -1,0 +1,48 @@
+export const pollInProgress: object = {};
+export const pollFinished: object = {
+  rNk4ePNVlJGwjCaVrjKFJbh3A0agpcZq6gYLfivOHYg: {
+    gas: 940,
+    result: {
+      status: 'success',
+      data: 'Write succeeded',
+    },
+    reqKey: 'rNk4ePNVlJGwjCaVrjKFJbh3A0agpcZq6gYLfivOHYg',
+    logs: 'GXRNaJAVVqBrZA05zJ1B66VRzos09z-s9M9LQAXdkvY',
+    events: [
+      {
+        params: [
+          'faucet-operation',
+          'k:db776793be0fcf8e76c75bdb35a36e67f298111dc6145c66693b0133192e2616',
+          9.4e-6,
+        ],
+        name: 'TRANSFER',
+        module: {
+          namespace: null,
+          name: 'coin',
+        },
+        moduleHash: 'rE7DU8jlQL9x_MPYuniZJf5ICBTAEHAIFQCB4blofP4',
+      },
+      {
+        params: [
+          'coin-faucet',
+          '938c8ec66283493844c205481f220cfbd7024bda5859c7897141f6a5c6fc52c1',
+          100,
+        ],
+        name: 'TRANSFER',
+        module: {
+          namespace: null,
+          name: 'coin',
+        },
+        moduleHash: 'rE7DU8jlQL9x_MPYuniZJf5ICBTAEHAIFQCB4blofP4',
+      },
+    ],
+    metaData: {
+      blockTime: 1696849672176432,
+      prevBlockHash: 'FxZHHgag5dS7f3hyNDVnTmeVOZZkdGtyvnqbkKspzVw',
+      blockHash: 'VnK4519k7MtweHWj9lENGQuIX_3plQj67LNAhgBljSc',
+      blockHeight: 3672752,
+    },
+    continuation: null,
+    txId: 3753319,
+  },
+};
