+++ packages/tools/e2e-tests/fixtures/mocks/send.ts
@@ -1,0 +1,3 @@
+export const send: object = {
+  requestKeys: ['rNk4ePNVlJGwjCaVrjKFJbh3A0agpcZq6gYLfivOHYg'],
+};
