+++ packages/tools/e2e-tests/page-objects/tools-app/components/aside.component.ts
@@ -1,0 +1,13 @@
+import type { Page } from '@playwright/test';
+
+export default class AsideComponent {
+  private _page: Page;
+
+  public constructor(page: Page) {
+    this._page = page;
+  }
+
+  public async navigateTo(ariaLabel: string): Promise<void> {
+    await this._page.getByRole('button', { name: ariaLabel }).click();
+  }
+}
