+++ packages/tools/e2e-tests/page-objects/tools-app/pages/account.page.ts
@@ -1,0 +1,9 @@
+import type { Page } from '@playwright/test';
+
+export default class AccountPage {
+  private readonly _page: Page;
+
+  public constructor(page: Page) {
+    this._page = page;
+  }
+}
