+++ packages/tools/e2e-tests/page-objects/tools-app/pages/home.page.ts
@@ -1,0 +1,13 @@
+import ToolsHeaderComponent from '../components/toolsHeader.component';
+
+import type { Page } from '@playwright/test';
+
+export default class HomePage {
+  private readonly _page: Page;
+  public header: ToolsHeaderComponent;
+
+  public constructor(page: Page) {
+    this._page = page;
+    this.header = new ToolsHeaderComponent(this._page);
+  }
+}
