+++ packages/tools/e2e-tests/page-objects/tools-app/pages/transactions.page.ts
@@ -1,0 +1,13 @@
+import AsideComponent from '../components/aside.component';
+
+import type { Page } from '@playwright/test';
+
+export default class TransactionsPage {
+  private readonly _page: Page;
+  public aside: AsideComponent;
+
+  public constructor(page: Page) {
+    this._page = page;
+    this.aside = new AsideComponent(this._page);
+  }
+}
