+++ packages/tools/e2e-tests/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "extends": "./node_modules/@kadena-dev/heft-rig/tsconfig-base.json",
+  "compilerOptions": {
+    "types": ["node"],
+    "rootDir": "./",
+    "target": "ES2020",
+    "skipLibCheck": true,
+    "noEmit": true,
+    "incremental": true,
+    "moduleResolution": "node",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "declaration": false, // fixes pnpm issue: https://github.com/microsoft/TypeScript/issues/42873
+    "declarationMap": false,
+    "jsx": "preserve",
+    "allowJs": true
+  },
+  "include": ["/**/*.ts", "/**/*.tsx"],
+  "exclude": ["node_modules"],
+}
